ContractLinkAI/backend/requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
dj-database-url==2.1.0
psycopg2-binary==2.9.9
whitenoise==6.6.0

# Celery & Redis
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-redis==5.4.0

# Web scraping
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0
playwright==1.41.1

# AI/OpenAI
openai==1.10.0

# Production server
gunicorn==21.2.0
uvicorn[standard]==0.27.0

# Utilities
python-dotenv==1.0.1
Pillow==10.2.0
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-asyncio==0.23.3
